API Allow declaring multiple domains
The MobileDomain and FullSiteDomain database columns now
accept a comma-separated list of values. This enables developers
to serve mobile content under multiple domains,
which is for example useful when used along the "subsites" module.
It is still recommended to have one canonical URL endpoint
for a unique content collection, though.

Conflicts:
	code/MobileSiteConfigExtension.php